Move Kino.Hub to Kino.Workspace
